Add assertion and logging to FindTaskCommand
